Add support for mv integration-test phase
- Added support to test the pf4j plugins in deployment mode within the SpringBootTest as integration tests
